refactor: move highlights/colors to galaxyline plugin
