+++ lua/cosmic/plugins/galaxyline/colors.lua
