+++ lua/cosmic/plugins/galaxyline/highlights.lua
